Changed install script

+++ install.sh
@@ -1,14 +1,26 @@
 #!/bin/bash
 # Script to install the configuration files
 
-mv ~/.config/i3/config ~/.config/i3/oldconfig
+if [! -f ~/.config/i3/oldconfig]; then
+    mv ~/.config/i3/config ~/.config/i3/oldconfig
+fi
+
 ln -s $PWD/i3/config ~/.config/i3/config
 
-mv ~/.bashrc ~/.bashrcold
+
+if [! -f ~/.bashrcold ]; then
+    mv ~/.bashrc ~/.bashrcold
+fi
+
 ln -s $PWD/.bashrc ~/.bashrc
 
-mv ~/.Xmodmap ~/.Xmodmapold
+if [! -f ~/.Xmodmapold ]; then
+    mv ~/.Xmodmap ~/.Xmodmapold
+fi
+
 ln -s $PWD/.Xmodmap ~/.Xmodmap
 
-mv ~/.Xresources ~/.Xresourcesold
+if [! -f ~/.Xresourcesold ]; then
+    mv ~/.Xresources ~/.Xresourcesold
+fi
 ln -s $PWD/.Xresources ~/.Xresources
